% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_options_position.R
\name{analyze_options_position}
\alias{analyze_options_position}
\title{Generate plot of options exposures}
\usage{
analyze_options_position(epics, positions)
}
\arguments{
\item{positions}{}
}
\value{
plotly object
}
\description{
Generate plot of options exposures
}
