% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_prices.R
\name{calc_strategy_prices}
\alias{calc_strategy_prices}
\title{Calculate strategy prices}
\usage{
calc_strategy_prices(common_prices, positions_matrix)
}
\arguments{
\item{positions_matrix}{}
}
\value{
Dataframe of strategy prices
}
\description{
Calculate strategy prices
}
