% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_by_time.R
\name{compute_implied_volatility}
\alias{compute_implied_volatility}
\title{Compute options implied volatility over time}
\usage{
compute_implied_volatility(option_leg, underlying_prices)
}
\arguments{
\item{option_leg}{Option leg}

\item{underlying_prices}{Close prices of the option's underlying}
}
\value{
Dataframe of implied_volatility
}
\description{
Compute options implied volatility over time
}
