% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_greeks_by_time.R
\name{compute_strategy_greeks}
\alias{compute_strategy_greeks}
\title{Compute options strategy greeks by time}
\usage{
compute_strategy_greeks(strategy)
}
\arguments{
\item{strategy}{Option strategy object}
}
\value{
Dataframe of greeks
}
\description{
Compute options strategy greeks by time
}
