% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_greeks_by_time.R
\name{compute_strategy_greeks_by_time}
\alias{compute_strategy_greeks_by_time}
\title{Compute options strategy greeks by time}
\usage{
compute_strategy_greeks_by_time(epics, resolution, n_prices, positions)
}
\arguments{
\item{epics}{List of instrument epics}

\item{resolution}{Resolution of prices}

\item{n_prices}{Number of prices}

\item{positions}{Vector containing position for each option in the strategy}
}
\description{
Compute options strategy greeks by time
}
