% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_prices.R
\name{compute_strategy_prices}
\alias{compute_strategy_prices}
\title{Compute strategy prices}
\usage{
compute_strategy_prices(strategy)
}
\arguments{
\item{strategy}{Strategy object}
}
\value{
Dataframe of strategy prices
}
\description{
Compute strategy prices
}
