% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ttm_years.R
\name{compute_ttm_years}
\alias{compute_ttm_years}
\title{Compute the time until maturity of an option, measured in years}
\usage{
compute_ttm_years(current_time, expiry)
}
\arguments{
\item{current_time}{Current DateTime}

\item{expiry}{DateTime of option expiry}
}
\value{
Time to maturity, measured in years
}
\description{
Compute the time until maturity of an option, measured in years
}
