% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_greeks_matrix.R
\name{create_greeks_matrix}
\alias{create_greeks_matrix}
\title{Create matrix of greeks for multiple options}
\usage{
create_greeks_matrix(epics)
}
\arguments{
\item{epics}{List of epics}
}
\description{
Create matrix of greeks for multiple options
}
