% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_market_env.R
\name{create_market_env}
\alias{create_market_env}
\title{Create market environment}
\usage{
create_market_env(node_id, pause = 0)
}
\arguments{
\item{node_id}{Integer ID of that particular market node}

\item{pause}{Seconds to pause between requests}
}
\value{

}
\description{
Create market environment
}
