% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_options_matrix.R
\name{create_options_matrix}
\alias{create_options_matrix}
\title{Create options matrix of prices from epics}
\usage{
create_options_matrix(epics, resolution, n_prices)
}
\arguments{
\item{epics}{List of epics}

\item{resolution}{Price resolution, eg. MINUTE_5, HOUR}

\item{n_prices}{Number of prices to request}
}
\value{
Matrix of dimension (n_prices, n_options)
}
\description{
Create options matrix of prices from epics
}
