% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{create_strategy}
\alias{create_strategy}
\title{Instantiate new strategy object}
\usage{
create_strategy(epics, positions, opening_prices = NULL, resolution, n_prices)
}
\arguments{
\item{epics}{Vector or epics for each leg of strategy}

\item{positions}{Integer vector specifiying positions in each leg}

\item{opening_prices}{Double vector of opening prices, only used when using a live strategy}

\item{resolution}{Plot resolution, eg. "HOUR", "MINUTE_5", "DAY"}

\item{n_prices}{Number of prices}
}
\value{

}
\description{
Instantiate new strategy object
}
