% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strategy_prices.R
\name{expand_prices}
\alias{expand_prices}
\title{Fill missing leg prices}
\usage{
expand_prices(leg_prices, union_datetime)
}
\arguments{
\item{leg_prices}{Dataframe of option leg prices}

\item{union_datetime}{Vector or datetimes to join on}
}
\value{
Prices dataframe with missing values filled with prior prices
}
\description{
Fill missing leg prices
}
