% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_current_greeks.R
\name{extract_current_greeks}
\alias{extract_current_greeks}
\title{Extract latest greeks}
\usage{
extract_current_greeks(greeks_df)
}
\arguments{
\item{greeks_df}{Dataframe of greeks over time}
}
\value{
Matrix of greeks
}
\description{
Extract latest greeks
}
