% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strategy_prices.R
\name{fill_prices}
\alias{fill_prices}
\title{Add additional data to option leg}
\usage{
fill_prices(legs)
}
\arguments{
\item{legs}{List of option legs with common expiry}
}
\value{
Option legs with new price data added based on union datetimes
}
\description{
Add additional data to option leg
}
