% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strategy_prices.R
\name{fill_strategy_prices}
\alias{fill_strategy_prices}
\title{Fill in missing prices for option legs}
\usage{
fill_strategy_prices(legs)
}
\arguments{
\item{legs}{List of option legs}
}
\value{
Option legs with new price data added
}
\description{
Fill in missing prices for option legs
}
