% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_positions.R
\name{get_current_positions}
\alias{get_current_positions}
\title{Get current positions}
\usage{
get_current_positions()
}
\value{
Summary dataframe of open positions
}
\description{
Get current positions
}
