% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greeks.R
\name{get_greeks}
\alias{get_greeks}
\title{Compute greeks for an option}
\usage{
get_greeks(epic, r = 0.05, b = 0, underlyer_annual_vol = 0.075,
  exposure = 1)
}
\arguments{
\item{epic}{Option epic}

\item{r}{Annualized rate of interest, \emph{as a decimal}, eg. 0.5 \% is 0.05}

\item{b}{Annualized cost of carry \emph{as a decimal}}

\item{underlyer_annual_vol}{Annualized volatility of the underlyer \emph{as a decimal}}

\item{exposure}{Amount bet per point, eg. 1}

\item{option_type}{"c" or "p", call or put}
}
\value{
Named list of option sensitivities/greeks: "delta", "gamma", "vega", "rho", "theta"
}
\description{
Compute greeks for an option
}
