% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_realized_volatility.R
\name{get_index_realized_volatility}
\alias{get_index_realized_volatility}
\title{Get index realized volatility from Oxford realized library}
\usage{
get_index_realized_volatility(index_symbol = ".SPX")
}
\arguments{
\item{index_symbol}{For complete list refer to Oxford Realized Volatility Library website}
}
\value{
Annualized Volatility from previous trading day
}
\description{
Get index realized volatility from Oxford realized library
}
