% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_option_details.R
\name{get_option_details}
\alias{get_option_details}
\title{Retrieve option details}
\usage{
get_option_details(epic)
}
\arguments{
\item{epic}{Option epic}
}
\value{
List of strike_price, option_type, expiry_datetime, underlyer
}
\description{
Retrieve option details
}
