% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_option_underlyer.R
\name{get_option_underlyer}
\alias{get_option_underlyer}
\title{Find option underlyer}
\usage{
get_option_underlyer(epic)
}
\arguments{
\item{epic}{Option epic}
}
\value{
Option underlyer
}
\description{
Find option underlyer
}
