% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startIGSession.R
\name{initiate_ig_session}
\alias{initiate_ig_session}
\title{Initiate IG Session}
\usage{
initiate_ig_session(env = "LIVE")
}
\arguments{
\item{env}{Which enviroment to use, one of "LIVE" or "DEMO"}
}
\value{
List of session information in global environment
}
\description{
Initiate IG Session
}
