% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_prices.R
\name{intersect_prices}
\alias{intersect_prices}
\title{Return commmon time series}
\usage{
intersect_prices(prices)
}
\arguments{
\item{prices}{List of price dataframes}
}
\value{
Input dataframes with common time series to all input dataframes
}
\description{
Return commmon time series
}
