% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ig_request.R
\name{make_ig_request}
\alias{make_ig_request}
\title{Make generic IG API request}
\usage{
make_ig_request(path, api_version, query = NULL)
}
\arguments{
\item{path}{API endpoint}

\item{api_version}{Certain endpoints are available for certain versions}

\item{query}{Optional query parameters}
}
\description{
Make generic IG API request
}
