% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceReq.R
\name{parse_prices}
\alias{parse_prices}
\title{Parse IG Index prices}
\usage{
parse_prices(prices_response, price_type)
}
\arguments{
\item{prices_response}{Response from IG index}

\item{price_type}{"Bid" or "ask"}
}
\value{
Dataframe of historical prices
}
\description{
Parse IG Index prices
}
