% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_by_time.R
\name{plot_implied_vols}
\alias{plot_implied_vols}
\title{Plot IV for each strategy leg}
\usage{
plot_implied_vols(strategy)
}
\arguments{
\item{strategy}{Strategy object}
}
\value{

}
\description{
Plot IV for each strategy leg
}
