% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positions_scenarios.R
\name{plot_positions_scenarios}
\alias{plot_positions_scenarios}
\title{Plot an existing positions PnL scenarios at a given time}
\usage{
plot_positions_scenarios(epics, opening_prices, directions, sizes,
  scenario_datetime)
}
\arguments{
\item{sizes}{}
}
\value{
3D surface plot of PnL scenarios at given time
}
\description{
Plot an existing positions PnL scenarios at a given time
}
