% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positions_scenarios.R
\name{plot_scenario_surface}
\alias{plot_scenario_surface}
\title{Plot option scenarios}
\usage{
plot_scenario_surface(scenario_matrix)
}
\arguments{
\item{scenario_matrix}{}
}
\description{
Plot option scenarios
}
