% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_option_pnl.R
\name{plot_strategy_pnl}
\alias{plot_strategy_pnl}
\title{Plot option Profit and Loss Scenarios}
\usage{
plot_strategy_pnl(strategy, plot_range = 100)
}
\arguments{
\item{strategy}{Strategy object}

\item{plot_range}{Maximum pnl range, integer}
}
\value{
Plot of PnL
}
\description{
Plot option Profit and Loss Scenarios
}
