% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strategy_prices.R
\name{plot_strategy_prices}
\alias{plot_strategy_prices}
\title{Generate plot of options exposures}
\usage{
plot_strategy_prices(strategy)
}
\arguments{
\item{strategy}{Strategy object}
}
\value{
plotly object
}
\description{
Generate plot of options exposures
}
