% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positions_scenarios.R
\name{plot_strategy_scenarios}
\alias{plot_strategy_scenarios}
\title{Plot an existing positions PnL scenarios at a given time}
\usage{
plot_strategy_scenarios(strategy, scenario_datetime)
}
\arguments{
\item{opening_prices}{}
}
\value{
3D surface plot of PnL scenarios at given time
}
\description{
Plot an existing positions PnL scenarios at a given time
}
