% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positions_scenarios.R
\name{plots_strategy_scenarios}
\alias{plots_strategy_scenarios}
\title{Plot strategy scenarios}
\usage{
plots_strategy_scenarios(strategy, scenario_datetime)
}
\arguments{
\item{strategy}{Strategy object}

\item{scenario_datetime}{}
}
\description{
Plot strategy scenarios
}
