% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_greeks_calculation.R
\name{prepare_greeks_calculation}
\alias{prepare_greeks_calculation}
\title{Extract option characteristics needed for greeks calculation}
\usage{
prepare_greeks_calculation(epic)
}
\arguments{
\item{epic}{Instrument epic}
}
\value{
List containing time to maturity, strike price, option type, underlyer price
}
\description{
Extract option characteristics needed for greeks calculation
}
