% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_strategy_prices.R
\name{refresh_strategy}
\alias{refresh_strategy}
\title{Refresh strategy prices to current datetime}
\usage{
refresh_strategy(strategy)
}
\arguments{
\item{strategy}{Option strategy}
}
\value{
Option strategy with up to date prices
}
\description{
Refresh strategy prices to current datetime
}
