% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_market_hierarchy.R
\name{request_market_navigation}
\alias{request_market_navigation}
\title{Request market node data}
\usage{
request_market_navigation(node_id)
}
\arguments{
\item{node_id}{Numeric id for a particular node in the market hierarchy}
}
\value{
Node data, containing either IDs of children nodes or market information if its a leaf
}
\description{
Request market node data
}
