% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceReq.R
\name{request_prices}
\alias{request_prices}
\title{Request historical prices}
\usage{
request_prices(epic = "IX.D.SPTRD.DAILY.IP", resolution = "HOUR",
  n_prices = 100)
}
\arguments{
\item{epic}{Instrument ID}

\item{resolution}{Time interval of prices eg. "HOUR", "DAY", "MINUTE_15"}

\item{n_prices}{Number of prices to request}
}
\value{
Dataframe of historical prices
}
\description{
Request historical prices
}
