% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceReq.R
\name{request_prices_range}
\alias{request_prices_range}
\title{Request prices for a given range}
\usage{
request_prices_range(epic, start_time, end_time, price_type = "bid")
}
\arguments{
\item{epic}{Instrument epic}

\item{start_time}{Datetime given as character in ISO8601 format UTC time, start period}

\item{end_time}{Datetime given as character in ISO8601 format UTC time, end period}

\item{price_type}{"Bid" or "Ask"}
}
\value{
Dataframe of prices
}
\description{
Request prices for a given range
}
