% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist_strategy.R
\name{save_strategy}
\alias{save_strategy}
\title{Save option strategy}
\usage{
save_strategy(strategy, path)
}
\arguments{
\item{strategy}{Strategy object}

\item{path}{File name}
}
\description{
Save option strategy
}
