% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_markets.R
\name{search_markets}
\alias{search_markets}
\title{Search IG Markets}
\usage{
search_markets(search_term)
}
\arguments{
\item{search_term}{}
}
\value{
Dataframe consisting of search results
}
\description{
Search IG Markets
}
