% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_plots.R
\name{style_plot}
\alias{style_plot}
\title{Apply common style to individual plots}
\usage{
style_plot(plot, y_axis_title = NULL, x_axis_title = NULL, show_legend = TRUE)
}
\arguments{
\item{plot}{Plotly object}

\item{axis_title}{Character}
}
\value{

}
\description{
Apply common style to individual plots
}
