% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_plots.R
\name{style_subplot}
\alias{style_subplot}
\title{Apply common style to plots}
\usage{
style_subplot(subplot, plot_title)
}
\arguments{
\item{subplot}{Plotly subplot object}

\item{title}{Character}
}
\value{

}
\description{
Apply common style to plots
}
