% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_market_hierarchy.R
\name{traverse_market_hierarchy}
\alias{traverse_market_hierarchy}
\title{Traverse IG markets hierarchy}
\usage{
traverse_market_hierarchy(node_id, pause = 0)
}
\arguments{
\item{node_id}{ID of the node from which to begin the traversal, see \href{https://labs.ig.com/rest-trading-api-reference/service-detail?id=544}{API documentation}}

\item{pause}{Due to api limits it may be necessary when traversing lots of nodes to introduce a pause to avoid exceeding said limits}
}
\value{
A dataframe containing markets details
}
\description{
Traverse IG markets hierarchy
}
