% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_option_pnl.R
\name{compute_option_pnl}
\alias{compute_option_pnl}
\title{Compute option PnL Scenarios at expiry}
\usage{
compute_option_pnl(option_leg, underlyer_prices)
}
\arguments{
\item{option_leg}{Option Leg object}

\item{underlyer_prices}{Vector of underlyer closing prices}
}
\value{

}
\description{
Compute option PnL Scenarios at expiry
}
