% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_greeks.R
\name{compute_strategy_greeks}
\alias{compute_strategy_greeks}
\title{Compute greeks for multiple options}
\usage{
compute_strategy_greeks(epics, position_matrix)
}
\arguments{
\item{epics}{List of epics}
}
\value{
Matrix of dimension (n_greeks, n_options)
}
\description{
Compute greeks for multiple options
}
