% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_greeks_by_time.R
\name{compute_strategy_greeks_by_time}
\alias{compute_strategy_greeks_by_time}
\title{Compute options strategy greeks by time}
\usage{
compute_strategy_greeks_by_time(strategy)
}
\arguments{
\item{strategy}{Option strategy object}

\item{resolution}{Resolution of prices}

\item{n_prices}{Number of prices}
}
\value{
Dataframe of greeks
}
\description{
Compute options strategy greeks by time
}
