% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_prices.R
\name{compute_strategy_prices}
\alias{compute_strategy_prices}
\title{Create options matrix of prices from epics}
\usage{
compute_strategy_prices(common_prices, positions_matrix)
}
\arguments{
\item{epics}{List of epics}

\item{resolution}{Price resolution, eg. MINUTE_5, HOUR}

\item{n_prices}{Number of prices to request}
}
\value{
Matrix of dimension (n_prices, n_options)
}
\description{
Create options matrix of prices from epics
}
