% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nodes_list.R
\name{create_nodes_list}
\alias{create_nodes_list}
\title{Create named list of market nodes}
\usage{
create_nodes_list(node_id, recursive = FALSE)
}
\arguments{
\item{node_id}{ID of the market node}
}
\value{
Named list where the node_ids constitute the values
}
\description{
Create named list of market nodes
}
