% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_strategy_greeks_by_time.R
\name{greeks_by_time}
\alias{greeks_by_time}
\title{Compute greeks over time for single option}
\usage{
greeks_by_time(option_leg, underlyer_prices, underlyer_datetimes)
}
\arguments{
\item{option_leg}{Option leg object}

\item{underlyer_prices}{Numeric vector containing option underlyer prices}

\item{underlyer_datetimes}{DateTime vector of the option underlyer prices}
}
\value{
List of option greeks for each underlyer price
}
\description{
Compute greeks over time for single option
}
