% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_by_time.R
\name{iv_by_time}
\alias{iv_by_time}
\title{Compute Implied Volatility over time}
\usage{
iv_by_time(underlying_prices, option_prices, option_datetimes, epic)
}
\arguments{
\item{underlying_prices}{Close prices of the option's underlying}

\item{option_prices}{Dataframe consisting of close prices and date time of the option}

\item{strike}{Option strike price}

\item{option_type}{Option type, "c" or "p"}
}
\description{
Compute Implied Volatility over time
}
