% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_option_pnl.R
\name{plot_strategy_pnl}
\alias{plot_strategy_pnl}
\title{Plot option Profit and Loss Scenarios}
\usage{
plot_strategy_pnl(strategy, underlyer_margin = 40)
}
\arguments{
\item{strategy}{Strategy object}

\item{underlyer_margin}{Integer specifying padding at margins of PnL graph}
}
\value{
Plot of PnL
}
\description{
Plot option Profit and Loss Scenarios
}
