% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strategy_prices.R
\name{plot_strategy_prices}
\alias{plot_strategy_prices}
\title{Generate plot of options exposures}
\usage{
plot_strategy_prices(epics, positions, resolution, n_prices)
}
\arguments{
\item{epics}{List of options epics}

\item{positions}{List of options positions, i.e. -1 for short one contract, +2 for long two contracts}

\item{resolution}{Plot resolution}

\item{n_prices}{Number of prices}
}
\value{
plotly object
}
\description{
Generate plot of options exposures
}
