% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strategy_scenario.R
\name{plot_strategy_scenario}
\alias{plot_strategy_scenario}
\title{Plot strategy scenarios}
\usage{
plot_strategy_scenario(strategy, scenario_datetime)
}
\arguments{
\item{scenario_datetime}{Datetime at which to evaluate scenarios}

\item{epics}{List of epics}
}
\description{
Plot strategy scenarios
}
